% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{barra_param_default}
\alias{barra_param_default}
\title{PARAMETER DEFAULT FUNCTIONS}
\usage{
barra_param_default()
}
\arguments{
\item{subset_reg}{NULL or sectorIDs vector, supporting more than one sectorID. 
If not NULL, regression would only be done within the sectorIDs, but the residuals would cover all the stocks.}

\item{h}{training windows.}

\item{tau}{half-life parameter.}

\item{lag}{lags for Newey-west adjustment.}

\item{shrink_q}{parameters for Beyasian shrink.}

\item{fix_itr_n}{parameters for Monte Carlo.}
}
\description{
PARAMETER DEFAULT FUNCTIONS
}
