% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{buildFactorList}
\alias{buildFactorList}
\alias{buildFactorList_lcfs}
\alias{buildFactorList_combi}
\title{buildFactorList}
\usage{
buildFactorList(factorFun = "gf_demo", factorPar = list(), factorDir = 1,
  factorRefine = refinePar_default("none"),
  factorName = default.factorName(factorFun, factorPar, factorDir),
  factorID = "", factorType = "", factorDesc = "")

buildFactorList_lcfs(factorID, factorRefine = refinePar_default("none"))

buildFactorList_combi(factorLists, wgts, factorDir = 1,
  factorRefine = refinePar_default("none"), factorName = "combi_factor",
  factorID = "", factorType = "", factorDesc = "")
}
\arguments{
\item{factorID}{a character string of factorID, available in table \code{CT_FactorLists()} .}
}
\value{
buildFactorList return a object of \bold{FactorList}, a list of parametres of factor setting

buildFactorList_lcfs get a object of \bold{FactorList} through "lcfs". See more detail in \code{\link{buildFactorLists_lcfs}}.
}
\description{
buildFactorList
}
\examples{
# -- set modelPar.factor by a factorlist
FactorList <- buildFactorList(factorFun="gf.pct_chg_per",factorPar=list(N=20))
modelPar <- modelPar.factor(modelPar,FactorList=FactorList)
# -- build a factorlist through "lcfs"
FactorList2 <- buildFactorList_lcfs(factorID="F000001")
}
