% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{cal_sigma_str}
\alias{cal_sigma_str}
\title{STRUCTURE SIGMA ESTIMATION FUNCTION}
\usage{
cal_sigma_str(sigma_ts_df, data_bank, barra_param)
}
\arguments{
\item{sigma_ts_df}{data table. The historical sigma data table, 
including columns 'date_end', 'stockID','sigma_ts','abs_res','gamma_blend'}

\item{data_bank}{data table. The data table obtained from load_data_bank(res_seri).}
}
\value{
data table, the structural sigma data table,
including columns 'date_end', 'stockID', 'sigma_str'
}
\description{
This function aimed to predict structural sigma using historical sigma.
}
