% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{check.rptDate}
\alias{check.rptDate}
\title{check.rptDate}
\usage{
check.rptDate(rptDate)
}
\arguments{
\item{rptDate}{a vector of rptDate , with class Date.}
}
\description{
Is it a proper rptDate object? If TRUE, return NULL; else return an error.
}
\examples{
check.rptDate(ymd(c(20100103,20141231,20130630)))
check.rptDate(ymd(c(20100630,20141231,20130930)))
}
\seealso{
\code{\link{rptDate.is}}
}
\author{
Ruifei.Yin
}
