% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{cov_mazi}
\alias{cov_mazi}
\title{Inner function for computing covariance}
\usage{
cov_mazi(f_mat, tau, lag)
}
\arguments{
\item{f_mat}{a matrix. factor return matrix. The matrix must not include date column, each column in f_mat is a factor return series.}

\item{tau}{integer number. The half-life parameter.}

\item{lag}{integer number. The Newey-west parameter.}
}
\description{
This function is a inner function to compute the Newey-west adjusted covariance matrix.
}
