% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{CT_industryList}
\alias{CT_industryList}
\title{CT_industryList}
\usage{
CT_industryList(std, level, ID, name, datasrc = defaultDataSRC())
}
\description{
CT_industryList
}
\examples{
re <- CT_industryList(std=c(9,3))
CT_industryList(std=9,level=1)
CT_industryList(ID=c("ES0310","ES0312"))
CT_industryList(name="SHENWAN")
}

