% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{CT_SecuCategory}
\alias{CT_SecuCategory}
\title{CT_SecuCategory}
\usage{
CT_SecuCategory(DM, MS)
}
\description{
CT_SecuCategory
}
\examples{
CT_SecuCategory(DM=c(1,4))
CT_SecuCategory(MS="xxx")
CT_SecuCategory()
}
