% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{CT_SecuMarket}
\alias{CT_SecuMarket}
\title{CT_SecuMarket}
\usage{
CT_SecuMarket(DM, MS)
}
\description{
CT_SecuMarket
}
\examples{
CT_SecuMarket(DM=c(83,90))
CT_SecuMarket(MS="xxx")
CT_SecuMarket()
}

