% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{defaultDataSRC}
\alias{defaultDataSRC}
\title{defaultDataSRC}
\usage{
defaultDataSRC()
}
\value{
a character string, the value of the default datasrc.
}
\description{
get the default datasrc. You can reset the default value by eg. \code{options(datasrc="quant")}
}
\examples{
# -- get the default datasrc
defaultDataSRC()
# -- reset 
options(datasrc="quant")
# -- reget
defaultDataSRC()
}
