% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_bcPower}
\alias{factor_bcPower}
\title{factor_bcPower}
\usage{
factor_bcPower(TSF, lambda = "auto", out_type = c("factor", "lambda",
  "both"), fname = "factorscore")
}
\arguments{
\item{lambda}{"auto" or a numeric.}

\item{out_type}{"factor", "lambda" or "both".}
}
\description{
factor_bcPower
}
\author{
Han.Qian
}
