% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_orthogan_name}
\alias{factor_orthogan_name}
\alias{factor_orthogon_single}
\alias{factor_orthogon}
\alias{factor_orthogon_transMat}
\title{factor orthogan}
\usage{
factor_orthogon_single(TSF, y, x, sectorAttr = defaultSectorAttr(),
  regType = c("lm", "glm"), res_adjust = FALSE)

factor_orthogon(TSF, method = c("symm", "reg"), forder, sectorAttr = NULL,
  regType = c("lm", "glm"))

factor_orthogon_transMat(TSF)
}
\examples{
tb_symm <- factor_orthogon(mtsf,method = "symm")
tb_reg <- factor_orthogon(mtsf,method = "reg")
MF.chart.Fct_corr(mtsf)
MF.chart.Fct_corr(tb_symm)
MF.chart.Fct_corr(tb_reg)
factor_orthogon_transMat(mtsf)
}
