% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_refine}
\alias{factor_refine}
\alias{factor_refine_MF}
\alias{MF.factor_refine}
\title{factor_refine}
\usage{
factor_refine(TSF, refinePar = refinePar_default(), drop_sector = TRUE)

factor_refine_MF(TSF, refinePar_lists, refinePar_names)

MF.factor_refine(mTSF, refinePar = refinePar_default(), drop_sector = TRUE)
}
\arguments{
\item{refinePar}{A refinePar list. Created by refinePar_default. Modified by setrefinePar.}

\item{refinePar_lists}{A list of (refinePar)s, each refinePar is a list built by refinePar_default.}

\item{refinePar_names}{The character vector of names, could be missing.}
}
\description{
factor_refine
}
\examples{
refinePar_lists <- list(refinePar_default(type = "none"),
                        refinePar_default(type = "reg"),
                        refinePar_default(type = "scale"))
mTSFR <- factor_refine_MF(rawTSFR, refinePar_lists)
MF.chart.Ngroup.spread(mTSFR)
}
\author{
Han.Qian
}
