% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{fill.indexquote000985}
\alias{fill.indexquote000985}
\title{fill.indexquote000985}
\usage{
fill.indexquote000985(windCode = "000985.CSI")
}
\description{
fill.indexquote000985
}
\examples{
library(WindR)
fill.indexquote000985()
}

