% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{fix_frtn_cov}
\alias{fix_frtn_cov}
\title{COV FIX FUNCTION FOR EIGENVALUES ADJUSTMENT.}
\usage{
fix_frtn_cov(cov_list, barra_param)
}
\arguments{
\item{cov_list}{A list. Each element in the list is a covariance matrix.}
}
\value{
A list. Each element in the list is a covariance matrix.
}
\description{
The function is aimed for eigenvalues adjustment for covariance matrix.
}
