% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{fix_frtn_cov2}
\alias{fix_frtn_cov2}
\title{COV FIX FUNCTION FOR BIAS ADJUSTMENT}
\usage{
fix_frtn_cov2(cov_list, frtn_seri, barra_param, result_type = c("cov",
  "bias"))
}
\arguments{
\item{cov_list}{A list. Each element in the list is a covariance matrix.}

\item{frtn_seri}{data table input. The factor return data table. The first column is date and each of the rest columns is a risk factor return series. The data is in day frequency.}

\item{result_type}{"cov" or "bias". 
If "cov", the output would be the cov_list after fixation. 
If "bias", the output would be the bias recording for fixation.}
}
\description{
The function is aimed for bias adjustment for covariance matrix.
}
