% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{fix_frtn_sigma}
\alias{fix_frtn_sigma}
\title{SIGMA FIX FUNCTION FOR BEYASIAN ADJUSTMENT}
\usage{
fix_frtn_sigma(sigma_i, barra_param)
}
\arguments{
\item{sigma_i}{data table, including columns 'date', 'stockID', 'sigma'}
}
\value{
sigma_i, same as input
}
\description{
This function is aimed for beyasian adjustment for sigma.
}
