% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{fix_frtn_sigma2}
\alias{fix_frtn_sigma2}
\title{SIGMA FIX FUNCTION FOR BIAS ADJUSTMENT}
\usage{
fix_frtn_sigma2(sigma_i, res_seri, barra_param, data_bank,
  result_type = c("sigma", "bias"))
}
\arguments{
\item{sigma_i}{data table, including columns 'date', 'stockID', 'sigma'}

\item{res_seri}{data table or data frame. The residuals data. 
The columns should be 'date', 'stockID' and 'res'.}

\item{data_bank}{data table. The data table obtained from load_data_bank(res_seri).}

\item{result_type}{"sigma" or "bias". 
If "sigma", the output would be the sigma_i after fixation. 
If "bias", the output would be the bias recording for fixation.}
}
\description{
The function is aimed for bias adjustment for sigma.
}
