% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{get_frtn_cov}
\alias{get_frtn_cov}
\title{COV COMPUTE FUNCTION}
\usage{
get_frtn_cov(frtn_seri, rebDates, barra_param)
}
\arguments{
\item{frtn_seri}{data table input. The factor return data table. 
The first column is date and each of the rest columns is a risk factor return series. The data is in day frequency.}

\item{rebDates}{vector input. The rebDates vector is the dates in which covariance matrix would be obtained.}
}
\value{
a list. Each element in the list is a covariance matrix. 
The name of each element is the rebDates. 
The matrix is computed base on the data no later than rebDate, and represents the ONE YEAR forward covariance matrix from that rebDate on.
}
\description{
COV COMPUTE FUNCTION
}
