% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{get_frtn_cov_single}
\alias{get_frtn_cov_single}
\title{COV COMPUTE FUNCTION FOR SINGLE PERIOD}
\usage{
get_frtn_cov_single(F_mat, barra_param)
}
\arguments{
\item{F_mat}{a matrix. factor return matrix. The matrix must not include date column, each column in F_mat is a factor return series.}
}
\value{
a matrix. The ONE DAY forward covariance matrix on the specific rebDate.
}
\description{
The function only works for computing one single covariance matrix.
}
