% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{get_frtn_sigma}
\alias{get_frtn_sigma}
\title{SIGMA COMPUTE FUNCTION}
\usage{
get_frtn_sigma(res_seri, rebDates, barra_param, data_bank)
}
\arguments{
\item{res_seri}{data table or data frame. The residuals data. 
The columns should be 'date', 'date_end', 'stockID' and 'res'.}

\item{rebDates}{vector input. The rebDates vector is the dates in which sigma would be obtained.}

\item{data_bank}{data table. The data table obtained from load_data_bank(res_seri).}
}
\value{
data table, including date, stockID, sigma. 
The sigma represents ONE YEAR forward sigma for each stock from that rebDate on.
}
\description{
SIGMA COMPUTE FUNCTION
}
