% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{get_frtn_sigma_single}
\alias{get_frtn_sigma_single}
\title{SIGMA COMPUTE FUNCTION FOR SINGLE PERIOD}
\usage{
get_frtn_sigma_single(res_seri_wide_sub, barra_param, data_bank)
}
\arguments{
\item{res_seri_wide_sub}{data table. the residual series data table in wide format, including columns date, date_end.}

\item{data_bank}{data table. The data table obtained from load_data_bank(res_seri).}
}
\value{
data table, including date, stockID, sigma. 
The sigma represents ONE DAY forward sigma for each stock from that rebDate on.
}
\description{
SIGMA COMPUTE FUNCTION FOR SINGLE PERIOD
}
