% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{gf_lcfs}
\alias{gf_lcfs}
\alias{gf.free_float_shares}
\alias{gf.free_float_sharesMV}
\title{gf_lcfs}
\usage{
gf_lcfs(TS, factorID)

gf.free_float_shares(TS)

gf.free_float_sharesMV(TS)
}
\description{
gf_lcfs
}
