% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk02_BARRA_LCDB.R
\name{getfrtn.pure_alpha}
\alias{getfrtn.pure_alpha}
\title{get frtn of pure factors}
\usage{
getfrtn.pure_alpha(begT, endT, result_type = c("fac_only", "all"))
}
\arguments{
\item{begT}{integar.}

\item{endT}{integar.}

\item{result_type}{fac_only or all, if all, result would contain c0 and sector rtn.}
}
\value{
daily return xts, ordered by date, each column is a pure factor.
}
\description{
get frtn of pure factors
}
\examples{
ggplot.WealthIndex(getfrtn.pure_alpha())
rtn.summary(getfrtn.pure_alpha(begT = 20180106, endT = 20180605))
}
