% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIFcontinuousCode}
\alias{getIFcontinuousCode}
\title{getIFcontinuousCode}
\usage{
getIFcontinuousCode(begT = as.Date("2010-04-16"), endT = Sys.Date())
}
\arguments{
\item{endT}{}
}
\value{
a datafame with cols:tradingday,IF00,IF01,IF02,IF03,IF04,firstday,lastday...
}
\description{
get the continuous index future code on every trading day
}
\examples{
begT <- as.Date("2010-04-16")
endT <- Sys.Date()
getIFcontinuousCode()
}
\author{
Ruifei.Yin
}

