% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIFlastfirstday}
\alias{getIFlastfirstday}
\title{getIFlastfirstday}
\usage{
getIFlastfirstday(begM = as.yearmon("2010-05"),
  endM = as.yearmon(Sys.Date()))
}
\arguments{
\item{begM}{beginning month,a yearmon object.}

\item{endM}{ending month,a yearmon object.}
}
\value{
a dataframe,with cols: date(of class Date),IF01(of class character) and firstday,lastday,openday(of class Date)
}
\description{
get the first,last and open dates of all IFs.
}
\details{
Note that the "firstday" is the date when the current-month IF switching,while the "openday" is the date the IF IPOing.
}
\examples{
getIFlastfirstday()
}
\author{
Ruifei.Yin
}

