% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIFlist}
\alias{getIFlist}
\title{getIFlist}
\usage{
getIFlist(begM = zoo::as.yearmon("2010-05"),
  endM = zoo::as.yearmon(Sys.Date()))
}
\arguments{
\item{begM}{beginning month,a yearmon object.}

\item{endM}{ending month,a yearmon object.}
}
\value{
a dataframe,with cols: date,IF01,IF02,IF03,IF04,firstday,lastday,openday
}
\description{
get the list(the 4 future at one time) of codes of all CSI300 index futures and the first and last day of every current-month future.
}
\examples{
getIFlist()
}
\author{
Ruifei.Yin
}

