% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIFrtn}
\alias{getIFrtn}
\title{getIFrtn}
\usage{
getIFrtn(code, begT, endT, adj = FALSE)
}
\arguments{
\item{code}{the code of the future(eg."IF00","IF01","IF02"...)}

\item{begT}{an object of class "Date"}

\item{endT}{an object of class "Date"}

\item{adj}{a logical,with default FALSE. if true, the daily return of future will be aligned with the trading time(starting from 9:30 and ending to 15:30) of CSI300 index.}
}
\value{
the daily return series of class xts.
}
\description{
get index future daily return series
}
\examples{
begT <- as.Date("2011-01-01")
endT <- as.Date("2011-02-01")
rtn <- getIFrtn("IF00",begT,endT)
rtn.adj <- getIFrtn("IF00",begT,endT,adj=TRUE)
}
\author{
Ruifei.Yin
}

