% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIndexComp}
\alias{getIndexComp}
\title{getIndexComp}
\usage{
getIndexComp(indexID, endT = Sys.Date(), drop = FALSE,
  datasrc = defaultDataSRC())
}
\arguments{
\item{indexID}{the stockID of the index}

\item{endT}{a vector of class \code{Date}. IF missing, then get the latest components.}

\item{drop}{if drop the field of date and return a vector when endT is length 1 ?}
}
\value{
If \code{endT} is missing or is length 1 and \code{drop} is TRUE, a vector of stockID of the components; else a dataframe, with cols: "date" and "stockID"
}
\description{
get the components of the specific index on certain day.
}
\examples{
tmp <- getIndexComp("EI000300",drop=TRUE) # get the latest components, a vector
tmp <- getIndexComp("EI000300") # get the latest components, a dataframe
tmp <- getIndexComp("EI000300",as.Date("2012-12-31")) # get the components on single day,a vector
tmp <- getIndexComp("EI000300",as.Date(c("2011-12-31","2012-12-31"))) # get the components on multi-days, a data frame
tmp <- getIndexComp("EI000300",as.Date(c("2005-12-31","2012-12-31"))) # get the components before index's pubdate, a data frame
}
