% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getIndexQuote}
\alias{getIndexQuote}
\title{getIndexQuote}
\usage{
getIndexQuote(indexs, begT = as.Date("1990-12-19"), endT = Sys.Date(),
  variables = select.list(CT_TechVars(datasrc = datasrc, secuCate = "EI",
  tableName = "QT_IndexQuote")[["varName"]], graphics = TRUE, multiple = TRUE),
  melt = FALSE, datasrc = defaultDataSRC())
}
\arguments{
\item{indexs}{a vector of charactor}

\item{begT}{an object of class "Date"}

\item{endT}{an object of class "Date"}

\item{variables}{a vector of charactor. Call funtion \code{CT_TechVars(secuCate="EQ")} to get all available variables of equity, funtion \code{CT_TechVars(secuCate="EI")} of equity index, and funtion \code{CT_TechVars(secuCate="IF")} of index future.}

\item{melt}{a logical. If FALSE(default), the style of result is "stockID+date~variable";If TRUE, the quote data will be melted(see the examples for details).}
}
\value{
the quote data, a data frame object with cols:stockID,date and the elements the param \code{variable} containing. If melt is TRUE, data frame with cols:stockID,date,variable,value
}
\description{
get Quote series of indexs in certain period
}
\examples{
indexs <- c("EI000001","EI000300")
begT <- as.Date("2015-01-01")
endT <- as.Date("2016-08-01")
variables <- c("close","pct_chg")
re <- getIndexQuote(indexs,begT,endT,variables,datasrc="local")
re <- getIndexQuote(indexs,begT,endT,variables,datasrc="jy")
}

