% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getPeriodrtn}
\alias{getPeriodrtn}
\title{getPeriodrtn}
\usage{
getPeriodrtn(SP, stockID, begT, endT, tradeType = c("close", "nextavg",
  "nextopen"), exclude.SP = FALSE, datasrc = defaultDataSRC())
}
\arguments{
\item{SP}{a \bold{SP} object ('stock*period'): a data frame with cols: "stockID","begT","endT"(with class of Date).}

\item{stockID}{a vector of stockID}

\item{begT}{a vector of Date}

\item{endT}{a vector of Date}

\item{tradeType}{a character string("close","nextavg","nextopen"),indicating the trading type.}

\item{exclude.SP}{a logical. If the \code{SP} shoud be inculded in the result?}
}
\value{
a data frame, with cols of SP and "periodrtn". Or a vector if \code{exclude.SP} is TRUE.
}
\description{
get the period return of the stocks
}
\note{
param SP and combination of stockID, begT and endT should  at least have one and only have one. The combination of vector stockID, begT and endT could be different length, which abide by the recycling rule.
}
\examples{
# -- with SP
getPeriodrtn(data.frame(stockID="EQ601313",begT=as.Date("2012-01-20"),endT=as.Date("2012-09-27")))
getPeriodrtn(data.frame(stockID="EQ601313",begT=as.Date("2012-01-20"),endT=as.Date("2012-09-27")),tradeType="nextopen")
TS <- getTS(getRebDates(as.Date('2007-03-17'),as.Date('2012-05-20')),'EI000300')
TS <- renameCol(TS,"date","begT")
SP <- data.frame(TS,endT=trday.offset(TS$begT,months(1)))
system.time(re <- getPeriodrtn(SP,datasrc="ts"))  # 8.40
system.time(re1 <- getPeriodrtn(SP,datasrc="local"))  # 1.94
system.time(re <- getPeriodrtn(SP,"nextavg",datasrc="ts")) # 10.80
system.time(re1 <- getPeriodrtn(SP,"nextavg",datasrc="local")) # 16.7

#-- with combination of vector stockID, begT and endT
getPeriodrtn(stockID="EQ000001",begT=as.Date("2012-01-01"), endT=as.Date("2012-01-01")+c(10,20,30))
getPeriodrtn(stockID=c("EQ000001","EQ000002","EQ000004"),begT=as.Date("2012-01-01"), endT=as.Date("2013-01-01"))
}
\author{
Ruifei.Yin
}

