% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getQuote}
\alias{getIFquote}
\alias{getIndexQuote}
\alias{getQuote}
\title{getQuote}
\usage{
getQuote(stocks, begT = as.Date("1990-12-19"), endT = Sys.Date(),
  variables = select.list(CT_TechVars(datasrc = datasrc, secuCate = "EQ",
  tableName = tableName)[["varName"]], graphics = TRUE, multiple = TRUE),
  melt = FALSE, split = if (length(stocks) > splitNbin) TRUE else FALSE,
  splitNbin = 300, tableName = "QT_DailyQuote",
  datasrc = defaultDataSRC())

getIndexQuote(stocks, begT = as.Date("1990-12-19"), endT = Sys.Date(),
  variables = select.list(CT_TechVars(datasrc = datasrc, secuCate = "EI",
  tableName = "QT_IndexQuote")[["varName"]], graphics = TRUE, multiple = TRUE),
  melt = FALSE, datasrc = defaultDataSRC())
}
\arguments{
\item{stocks}{a charactor vector of stockID or indexID.}

\item{begT}{an object of class "Date"}

\item{endT}{an object of class "Date"}

\item{variables}{a vector of charactor. Call funtion \code{CT_TechVars(secuCate="EQ")} to get all available variables of equity, funtion \code{CT_TechVars(secuCate="EI")} of equity index, and funtion \code{CT_TechVars(secuCate="IF")} of index future.}

\item{melt}{a logical. If FALSE(default), the style of result is "stockID+date~variable";If TRUE, the quote data will be melted(see the examples for details).}

\item{tableName}{character string, giving the table name from which get data.}
}
\value{
the quote data, a data frame object with cols:stockID,date and the elements the param \code{variable} containing. If melt is TRUE, data frame with cols:stockID,date,variable,value
}
\description{
get Quote series of stocks, indexs, futures in certain period
}
\examples{
# get stock's quote
stocks <- c("EQ002001","EQ002002")
begT <- as.Date("2007-01-01")
endT <- as.Date("2013-02-01")
variables <- c("close","pct_chg","vwap")
re <- getQuote(stocks,begT,endT,variables,datasrc="local")
re <- getQuote(stocks,begT,endT,variables,melt=TRUE,datasrc="local")
stocks <- getIndexComp("EI000300")
system.time(uu2 <- getQuote(stocks,begT,endT,variables,datasrc="local",split=FALSE))
system.time(uu1 <- getQuote(stocks,begT,endT,variables,datasrc="local",split=TRUE))
# get index quote
indexs <- c("EI000001","EI000300")
begT <- as.Date("2015-01-01")
endT <- as.Date("2016-08-01")
variables <- c("close","pct_chg")
re <- getIndexQuote(indexs,begT,endT,variables,datasrc="local")
re <- getIndexQuote(indexs,begT,endT,variables,datasrc="jy")
}
\author{
Ruifei.Yin
}

