% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getQuote_ts}
\alias{getQuote_ts}
\title{getQuote_ts}
\usage{
getQuote_ts(stocks, begT, endT, variables, Cycle = "cy_day()", Rate = 0,
  RateDay = 0, melt = FALSE, split = if (length(stocks) > splitNbin) TRUE
  else FALSE, splitNbin = 50)
}
\arguments{
\item{stocks}{a vector of charactor}

\item{begT}{an object of class "Date"}

\item{endT}{an object of class "Date"}

\item{variables}{a vector of charactor,elements of which could be stockID,stockName,date,price,open,high,low,vol,amount,yclose,sectional_yclose,cjbs,...}

\item{Cycle}{a charactor string,eg."cy_day()","cy_30m()","cy_month()",...}

\item{Rate}{a integer,giving the type of rights adjustment, could be one of 0(no adjustment),1(geometric adjustment),2(simple adjustment),3}

\item{RateDay}{a integer,giving the base date of right adjustment,could be one of 0(the last trading day),-1(the IPO date),or a tinysoft date integer(eg.\code{rdate2ts(as.Date("2010-01-02"))})}

\item{melt}{a logical. If FALSE(default), the style of result is "stockID+date~variable";If TRUE, the quote data will be melted(see the examples for details).}
}
\value{
the quote data, a data frame object with cols:stockID,stockName,date(\bold{of class POSIXct}) and the elements the param \code{variable} containing
}
\description{
get Quote series of some stocks in certain period
}
\examples{
stocks <- c("SZ002001","SZ002002")
begT <- as.Date("2011-01-01")
endT <- as.Date("2011-02-01")
variables <- c("price","open")
qt.asis <- getQuote_ts(stocks,begT,endT,variables)
qt.melt <- getQuote_ts(stocks,begT,endT,variables,melt=TRUE)
}
\author{
Ruifei.Yin
}
