% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getrptDate_newest}
\alias{getrptDate_newest}
\title{getrptDate_newest}
\usage{
getrptDate_newest(TS, stockID, endT = Sys.Date(), freq = c("q", "y", "h"),
  mult = c("last", "first", "all"), drop = FALSE,
  datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{a \bold{TS} object}

\item{stockID}{a vector of stockID}

\item{endT}{a vector of Date}

\item{mult}{a character string. Could be one of "last","first","all". IF a listed company publish more than one financial reports in a single day, which one should be returned? the newest one(the default value), the earliest one or all of them? See example for dedail.}

\item{drop}{a logical. Shoud the \code{TS} be exculded in the result?}
}
\value{
a data.frame,with the same cols of TS,added by "\code{rptDate}". Or a vector if \code{drop} is TRUE.
}
\description{
get the newest rptDate  of the stocks on specific dates.
}
\note{
param TS and combination of stockID and endT should  at least have one and only have one. The combination of vector stockID and endT could be different length, which abide by the recycling rule.
}
\examples{
# - with TS
TS <- getTS(getRebDates(as.Date('2007-03-17'),as.Date('2012-05-20'),rebFreq="year"),'EI000300')
getrptDate_newest(TS)
# - one stock, multiple dates
getrptDate_newest(stockID="EQ000001", endT=seq(from=as.Date("2010-12-31"),to=as.Date("2011-12-31"),by="month"))
# - one date, multiple stocks
getrptDate_newest(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2003-04-30"))
# - multi="all"
getrptDate_newest(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2003-04-30"),mult="all")
# - multi="first"
getrptDate_newest(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2003-04-30"),mult="first")
}
\author{
Ruifei.Yin
}
