% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getrptTS}
\alias{getrptTS}
\title{getrptTS}
\usage{
getrptTS(begT, endT, freq = c("q", "y", "h"), dir = 0, univ, rptDates,
  stocks)
}
\arguments{
\item{freq}{"q","y","h"}

\item{dir}{1L,-1L,or 0L}

\item{univ}{a index or sector ID.}
}
\value{
a \bold{rptTS} object.a dataframe,with cols:
  \itemize{
  \item date: the rptDates
  \item stockID: the stockID 
  }.IF rptDates or stocks is not missing then use them directly, else get rptDates and stocks by \code{\link{rptDate.get}} and \code{\link{getComps}}.
}
\description{
getrptTS
}
\examples{
re1 <- getrptTS(begT=as.Date("2011-02-06"),endT=as.Date("2011-10-23"),univ="EI000300")
re2 <- getrptTS(rptDates=as.Date(c("2016-03-31","2016-09-30")),univ="EI000300")
re3 <- getrptTS(rptDates=as.Date(c("2016-03-31","2016-09-30")),stocks=c("EQ000001","EQ000002"))
re4 <- getrptTS(begT=as.Date("2011-02-06"),endT=as.Date("2011-10-23"),stocks=c("EQ000001","EQ000002"))
}
\author{
Ruifei.Yin
}
