% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getSectorComp}
\alias{getSectorComp}
\title{getSectorComp}
\usage{
getSectorComp(sectorID, endT = Sys.Date(), drop = TRUE,
  datasrc = defaultDataSRC())
}
\arguments{
\item{sectorID}{ID of the sector. eg. "ES09440000" for "financial service" the 1st level sector of SHENWAN; "ES09440100" for "bank"  the 2nd level sector of SHENWAN.  Get details with \code{CT_industryList()}}

\item{endT}{a vector of class \code{Date}. IF missing, then get the latest components.}

\item{drop}{if drop the field of date and return a vector when endT is length 1 ?}
}
\value{
If \code{endT} is missing or is length 1 and \code{drop} is TRUE, a vector of stockID of the components; else a dataframe, with cols: "date" and "stockID"
}
\description{
get the components of the specific sector on certain day.
}
\examples{
tmp <- getSectorComp("ES09440000") # get the latest components, a vector
tmp <- getSectorComp("ES09440000",as.Date("2012-12-31")) # get the components on single day,a vector
tmp <- getSectorComp("ES09440000",as.Date(c("2011-12-31","2012-12-31"))) # get the components on multi-days, a data frame
}
\seealso{
Other getComps functions: \code{\link{getComps}},
  \code{\link{getIndexCompWgt}}
}

