% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getSectorID}
\alias{getSectorID}
\title{getSectorID}
\usage{
getSectorID(TS, stockID, endT = Sys.Date(),
  sectorAttr = defaultSectorAttr(), ret = c("ID", "name"), drop = FALSE,
  datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{a \bold{TS} object}

\item{stockID}{a vector of stockID}

\item{endT}{a vector of Date}

\item{sectorAttr}{a list with two elements: \code{std}, a integer ID number, indicating the sector standard(9 for SHENWAN sector,3 for ZHONGXIN sector.See more by \code{\link{CT_sectorSTD}}; \code{level}, a integer, giving the level of sector,the default is 1.}

\item{ret}{a charactor string,could be one of "ID" or "name",indicating sectorID or sectorName returned.}

\item{drop}{a logical. Shoud the \code{TS} be exculded in the result?}
}
\value{
a data.frame,with the same cols of TS,added by "\code{sector}". Or a vector if \code{drop} is TRUE. You can get more sector infomation by \code{\link{CT_industryList}}
}
\description{
get the sectorID of the stocks on specific dates.
}
\note{
param TS and combination of stockID and endT should  at least have one and only have one. The combination of vector stockID and endT could be different length, which abide by the recycling rule.
}
\examples{
# - with TS
TS <- getTS(getRebDates(as.Date('2007-03-17'),as.Date('2012-05-20')),'EI000300')
getSectorID(TS)
# - one stock, multiple dates
getSectorID(stockID="EQ000001", endT=as.Date(c("2010-01-01","2012-01-01","2013-01-01")))
# - one date, multiple stocks
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"))
# - get 'ZHONGXIN' sector
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"), sectorAttr=list(3,1), ret="name")
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"), sectorAttr=list(3,2), ret="name")
# -- combined sectorAttr
factorList1 <- buildFactorList(factorFun = "gf.mkt_cap", factorStd = "norm", factorNA = "na")
test <- getSectorID(TS, sectorAttr= list(std=list(factorList1,33),level=list(5,1)))
}
\author{
Ruifei.Yin
}

