% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getTech}
\alias{getTech}
\alias{TS.getTech}
\title{getTech}
\usage{
getTech(TS, variables = select.list(CT_TechVars(datasrc = datasrc, secuCate =
  "EQ", tableName = tableName)[["varName"]], graphics = TRUE, multiple = TRUE),
  tableName = "QT_DailyQuote", datelist, stockID, drop,
  datasrc = defaultDataSRC())

TS.getTech(TS, variables = select.list(CT_TechVars(datasrc = datasrc, secuCate
  = "EQ", tableName = tableName)[["varName"]], graphics = TRUE, multiple =
  TRUE), tableName = "QT_DailyQuote", datasrc = defaultDataSRC())
}
\description{
getTech
}
