% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getTradeList}
\alias{getTradeList}
\title{getTradeList}
\usage{
getTradeList(port_ini, port_obj, money_obj, splitN = 1,
  outputstyle = c("hs"))
}
\arguments{
\item{port_ini}{a dataframe with cols: "stockID" (with EQ format),"amount".}

\item{port_obj}{a dataframe with cols: "stockID","wgt"}

\item{money_obj}{a numeric}

\item{splitN}{a integer}
}
\value{
a list with 3 items: tradeList_N,tradeList_remain,summaryTable
}
\description{
getTradeList
}
\author{
Han.Qian
}
