% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{gf_cap}
\alias{gf_cap}
\title{gf_cap}
\usage{
gf_cap(TS, log = FALSE, bc_lambda = NULL, var = c("float_cap", "mkt_cap",
  "free_cap"), na_fill = TRUE, varname = "factorscore",
  datasrc = "memory")
}
\arguments{
\item{bc_lambda}{NULL, "auto", or a specific numeric.}
}
\description{
gf_cap
}
\examples{
system.time(re <- gf_cap(ts,datasrc = "local"))
system.time(re2 <- gf_cap(ts,datasrc = "memory"))
}
