% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{healthy_sample_test}
\alias{healthy_sample_test}
\title{HEALTHY SAMPLE TEST}
\usage{
healthy_sample_test(vec)
}
\value{
numeric vector. The values equal or bigger than 1 representing unhealthy sample points.
}
\description{
HEALTHY SAMPLE TEST
}
