% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{is_component}
\alias{is_component}
\title{is component of specific sector or index?}
\usage{
is_component(TS, stockID, endT = Sys.Date(), sectorID, drop = FALSE,
  datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{}

\item{stockID}{}

\item{endT}{}

\item{sectorID}{a character string. The ID of the index, sector or plate. See detail in \code{\link{getComps}}}

\item{drop}{}

\item{datasrc}{}
}
\value{
a dataframe or a vector(if drop==TRUE)
}
\description{
is component of specific sector or index?
}
\examples{
is_component(stockID = c("EQ000001","EQ300089"),sectorID = "EI000300",drop=TRUE)
is_component(stockID = c("EQ000001","EQ300089"),sectorID = "ES09440000")
TS <- getTS(as.Date(c("2014-01-01","2016-01-01")),indexID = "EI000906")
re <- is_component(TS,sectorID = "EI000300")
}
