% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{is_delist}
\alias{is_delist}
\title{is_delist}
\usage{
is_delist(TS, nearby = months(2), datelist, stockID, drop = FALSE,
  datasrc = "jy")
}
\description{
judeg whether stocks to be delisted
}
