% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{is_priceLimit}
\alias{is_priceLimit}
\title{is_priceLimit}
\usage{
is_priceLimit(TS, nearby = 0, lim = c(-10, 10), datelist, stockID, drop,
  datasrc = defaultDataSRC())
}
\arguments{
\item{nearby}{a integer vector. default is 1L, which means remove the suspending of the next tradingday. see detail in \code{\link{trday.nearby}}.}

\item{lim}{a vector of length 2.}
}
\description{
is_priceLimit
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- is_priceLimit(TS)
re1 <- is_priceLimit(TS,nearby=-1:1)
re2 <- is_priceLimit(TS,lim=c(-Inf,10)) #  limit-up
}
\author{
Ruifei.yin
}

