% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{is_st}
\alias{is_st}
\title{is_st}
\usage{
is_st(TS, datelist, stockID, drop = FALSE, datasrc = defaultDataSRC())
}
\description{
is st stock
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- is_st(TS)
}
\author{
Ruifei.yin
}
