% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R,
%   R/pub02_dataGettingFuncs.R
\name{is_suspend}
\alias{is_suspend}
\alias{is_suspend}
\title{is_suspend}
\usage{
is_suspend(TS, nearby = 0, datelist, stockID, drop,
  datasrc = defaultDataSRC())

is_suspend(TS, nearby = 0, datelist, stockID, drop,
  datasrc = defaultDataSRC())
}
\arguments{
\item{nearby}{a integer vector. 0 means today, 1 means next tradingday. default is 0L. see detail in \code{\link{trday.nearby}}.}

\item{nearby}{a integer vector. 0 means today, 1 means next tradingday. default is 0L. see detail in \code{\link{trday.nearby}}.}
}
\description{
is_suspend

is_suspend
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- is_suspend(TS) #  Suspend of nextday
re1 <- is_suspend(TS,nearby=c(0,1))#  Suspend of today and nextday
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- is_suspend(TS) #  Suspend of nextday
re1 <- is_suspend(TS,nearby=c(0,1))#  Suspend of today and nextday
}
\author{
Ruifei.yin

Ruifei.yin
}
