% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.add.LC_IndexComponentsWeight}
\alias{lcdb.add.LC_IndexComponentsWeight}
\title{add new security to local database.}
\usage{
lcdb.add.LC_IndexComponentsWeight(IndexID)
}
\description{
add new security to local database.
}
\examples{
lcdb.add.LC_IndexComponentsWeight(c("EI000905","EI000906"))
}

