% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.build.QT_FreeShares}
\alias{lcdb.build.QT_FreeShares}
\title{lcdb.build.QT_FreeShares}
\usage{
lcdb.build.QT_FreeShares(filename = "D:/sqlitedb/FREE_SHARES.csv")
}
\description{
lcdb.build.QT_FreeShares
}
\examples{
lcdb.build.QT_FreeShares(filename="D:/sqlitedb/FREE_SHARES.csv")
}

