% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.export2csv}
\alias{lcdb.export2csv}
\title{lcdb.export2csv}
\usage{
lcdb.export2csv(dbname, tablename, path = "D:/sqlitedb")
}
\description{
lcdb.export2csv
}
\examples{
lcdb.export2csv("main","CT_TechVars")
lcdb.export2csv("main","QT_FreeShares")
lcdb.export2csv("fs","CT_FactorLists")
lcdb.export2csv("fs_r","CT_FactorLists_r")
}
