% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.fill.indexquote000985}
\alias{lcdb.fill.indexquote000985}
\title{lcdb.fill.indexquote000985}
\usage{
lcdb.fill.indexquote000985(windCode = "000985.CSI")
}
\description{
lcdb.fill.indexquote000985
}
\examples{
library(WindR)
lcdb.fill.indexquote000985()
}

