% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.init.indexquote000985}
\alias{lcdb.init.indexquote000985}
\title{lcdb.init.indexquote000985}
\usage{
lcdb.init.indexquote000985(windCode = "000985.CSI")
}
\description{
only need to run ONE TIME. filling the nearly quote data of EI000985.
}
\examples{
library(WindR)
lcdb.init.indexquote000985()
}

