% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.init.IndexQuote_000985}
\alias{lcdb.init.IndexQuote_000985}
\title{lcdb.init.IndexQuote_000985}
\usage{
lcdb.init.IndexQuote_000985(begT = 19900101)
}
\description{
initiate the quote data of index 000985 through Wind API.
}
\examples{
lcdb.init.IndexQuote_000985()
}
