% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.init.IndexQuote_000985E}
\alias{lcdb.init.IndexQuote_000985E}
\alias{lcdb.init.IndexQuote_000985}
\alias{lcdb.init.QT_FreeShares}
\alias{lcdb.init.CT_TechVars}
\alias{lcdb.init_main}
\alias{lcdb.init_qt}
\alias{lcdb.init.QT_DailyQuote}
\alias{lcdb.init.QT_DailyQuote2}
\alias{lcdb.init.CT_FactorLists}
\alias{lcdb.init.QT_FactorScore}
\alias{lcdb.init_fs}
\title{lcdb.init}
\usage{
lcdb.init.IndexQuote_000985E(begT = 20050101)

lcdb.init.IndexQuote_000985(begT = 19900101)

lcdb.init.QT_FreeShares(filename = "D:/sqlitedb/QT_FreeShares.csv")

lcdb.init.CT_TechVars(filename = "D:/sqlitedb/CT_TechVars.csv")

lcdb.init_main(begT = 19900101, endT = 99990101)

lcdb.init_qt()

lcdb.init.QT_DailyQuote(begT = 19900101, endT = 99990101)

lcdb.init.QT_DailyQuote2(begT = 19900101, endT = 99990101)

lcdb.init.CT_FactorLists(filename = "D:/sqlitedb/CT_FactorLists.csv")

lcdb.init.QT_FactorScore(begT = 20050104, endT = 99990101)

lcdb.init_fs()
}
\description{
initialize all the tables in sqlitdb
}
\examples{
# A correct throughout process of initialize the whole local database:
# 1. expert 3 csv tables:
lcdb.export2csv("main","CT_TechVars")
lcdb.export2csv("main","QT_FreeShares")
lcdb.export2csv("fs","CT_FactorLists")
# 2. build 3 empty sqlite files: qt.db, fs.db, main.db
# 3. initialize 3 database files, in proper order:
lcdb.init_qt()
lcdb.init_main()
lcdb.init_fs()
}
