% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.update.IndexQuote_000985E}
\alias{lcdb.update.IndexQuote_000985E}
\title{lcdb.update.IndexQuote_000985E}
\usage{
lcdb.update.IndexQuote_000985E(begT, endT)
}
\description{
lcdb.update.IndexQuote_000985E
}
\examples{
lcdb.update.IndexQuote_000985E()
}
