% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.update.QT_FactorScore}
\alias{lcdb.update.QT_FactorScore}
\title{lcdb.update.QT_FactorScore}
\usage{
lcdb.update.QT_FactorScore(begT, endT, stockID, loopFreq = "month",
  type = c("alpha", "risk"))
}
\arguments{
\item{begT}{the begin date of the updating}

\item{endT}{the end date of the updating}

\item{stockID}{a vector of stockID}
}
\description{
update \bold{all} factorscores in table CT_FactorLists (see \code{CT_FactorLists()}).
}
\examples{
# update factors on certain time
lcdb.update.QT_FactorScore(20130322,20130330)
# update factors of certain stocks
lcdb.update.QT_FactorScore(20130322,20130330,c("EQ000001","EQ000002")) 
# update factors on certin time, of certain stocks
lcdb.update.QT_FactorScore(20130322,20130330,c("EQ000001","EQ000002"))
}
\seealso{
\code{\link{lcfs.update}}
}
