% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.update.QT_FreeShares}
\alias{lcdb.update.QT_FreeShares}
\title{lcdb.update.QT_FreeShares}
\usage{
lcdb.update.QT_FreeShares(begT, endT, Freq = "week")
}
\description{
update QT_FreeShares through Wind API.
}
\examples{
lcdb.update.QT_FreeShares()
}
