% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.update}
\alias{lcdb.fix.swindustry}
\alias{lcdb.update}
\alias{lcdb.update.CT_IndustryList}
\alias{lcdb.update.CT_SystemConst}
\alias{lcdb.update.LC_ExgIndustry}
\alias{lcdb.update.LC_IndexComponent}
\alias{lcdb.update.LC_IndexComponentsWeight}
\alias{lcdb.update.LC_PerformanceGrowth}
\alias{lcdb.update.LC_RptDate}
\alias{lcdb.update.QT_DailyQuote}
\alias{lcdb.update.QT_DailyQuote2}
\alias{lcdb.update.QT_IndexQuote}
\alias{lcdb.update.QT_TradingDay}
\alias{lcdb.update.QT_UnTradingDay}
\alias{lcdb.update.SecuMain}
\alias{lcfs.update}
\title{update the local database}
\usage{
lcdb.update()

lcdb.update.SecuMain()

lcdb.update.QT_TradingDay()

lcdb.update.CT_SystemConst()

lcdb.update.CT_IndustryList()

lcdb.update.LC_ExgIndustry()

lcdb.update.LC_IndexComponent()

lcdb.update.LC_IndexComponentsWeight()

lcdb.update.QT_UnTradingDay()

lcdb.update.QT_IndexQuote()

lcdb.update.QT_DailyQuote()

lcdb.update.QT_DailyQuote2()

lcdb.update.LC_RptDate()

lcdb.update.LC_PerformanceGrowth()

lcdb.fix.swindustry()

lcfs.update(endT = Sys.Date())
}
\value{
nothing.
}
\description{
update the local database

Due to the SHENWAN inustry standard changing in 2014, fix local database's shenwan industry rule's bug and make the rule keep consistent. The new industry standard is 33.  Update 3 related local tables:CT_SystemConst,CT_IndustryList and LC_ExgIndusry.
}
\examples{
lcdb.fix.swindustry()
}
\author{
Andrew Dow
}

