% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.updatetime}
\alias{lcdb.updatetime}
\title{lcdb.updatetime}
\usage{
lcdb.updatetime()
}
\value{
a datafrme, with cols: "table", "updatetime".
}
\description{
get the updatetime of tables in lcdb
}
