% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcfs.add}
\alias{lcfs.add}
\title{lcfs.add}
\usage{
lcfs.add(factorFun, factorPar = "", factorDir, factorID,
  factorName = default.factorName_inner(factorFun, factorPar, factorDir),
  factorType = "", factorDesc = "", begT = as.Date("1990-01-01"),
  endT = Sys.Date(), splitNbin = "month")
}
\arguments{
\item{factorFun}{a character string naming the function to get the factor scores}

\item{factorPar}{a character string, containing the parameters of the \code{factorFun}. Note that unlike in \code{\link{getTSF}}, here the factorPar could not be a list, because it need to be written into database.}

\item{factorDir}{a integer,should be 1 or -1 (1 for the positive factor,-1 for the negative one). \bold{Note that} the \code{factorDir} here is only used to write a record into table \code{"CT_FactorLists"}, not used when getting \code{TSF}. So that the factorscore in table \code{"QT_DailyQuote"} is kept \bold{"raw", without adding the dirrection infomation}.}

\item{factorID}{a character string}

\item{factorName}{a character string. IF missing, then take a default name by function \code{default.factorName}.}

\item{factorType}{a character string}

\item{factorDesc}{a character string}

\item{begT}{the begin date of the adding/updating}

\item{endT}{the end date of the adding/updating}

\item{splitNbin}{a character of interval specification(see \code{\link{cut.Date}} for detail). Specify the time interval when looping of getting the \code{TSF} object.}
}
\value{
Write data into the local sqlite database, returning NULL.
}
\description{
add/update a factorscore column in local sqlite table \code{"QT_DailyQuote"}. On the same time, correspondingly, add/update a record into table \code{"CT_FactorLists"} and table \code{"CT_TechVars"}.
}
\examples{
system.time(lcfs.add(factorFun="gf.F_rank_chg",factorPar="lag=60,con_type=\\"1,2\\"", factorDir=1, factorID="F000999"))
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link{getTSF}},\code{\link{modelPar.factor}}, \code{\link{lcfs.update}}
}

