% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcfs.update}
\alias{lcfs.update}
\title{lcfs.update}
\usage{
lcfs.update(factorID, begT = 19900101, endT = 99999999, stockID,
  splitNbin = "month")
}
\arguments{
\item{begT}{the begin date of the updating}

\item{endT}{the end date of the updating}
}
\description{
lcfs.update
}
\examples{
lcfs.update("F000008",20130322,20130330,c("EQ000001","EQ000002"))
}
