% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{load_data_bank}
\alias{load_data_bank}
\title{DATA_BANK LOAD FUNCTION}
\usage{
load_data_bank(res_seri)
}
\arguments{
\item{res_seri}{data table or data frame. The residuals data. 
The columns should be 'date', 'date_end', 'stockID' and 'res'.}
}
\description{
This function is used to prepare the data for get_frtn_sigma() and fix_frtn_sigma2().
It simply get factor scores from local data base for regression usage in further procedures.
}
