% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{memory.load}
\alias{memory.load}
\title{memory.load}
\usage{
memory.load(reload = FALSE)
}
\arguments{
\item{reload}{a logic. If reload the memory data?}
}
\value{
return NULL, but load some data into memory.
}
\description{
load frequently used data into memory, including \code{.tradingdays}, \code{.QT_sus_res}, ...
}
\author{
Ruifei.Yin
}

