% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{MF_Turnover_annual}
\alias{MF_Turnover_annual}
\title{MF_Turnover_annual}
\usage{
MF_Turnover_annual(fundID, begrptDate, endrptDate)
}
\description{
MF_Turnover_annual
}
\examples{
begrptDate <- as.Date("2015-06-30")
endrptDate <- as.Date("2016-12-31")
fundID <- c("519983.OF","519987.OF")
MF_Turnover_annual(fundID, begrptDate, endrptDate)
}
