% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub04_Fund_DataGettingFuncs.R
\name{MF_getTF}
\alias{MF_getTF}
\alias{MF_TF_reshape}
\alias{MF_CT_nonalphafund}
\alias{MF_getAssetPort}
\alias{MF_getTFSW}
\title{MF_getTF}
\usage{
MF_getTF(RebDates, fundtype = c("equity", "aggr_allo", "index",
  "indexenhance"), rm = c("namewithC", "initperiod"), N = 90)

MF_TF_reshape(TF, type = c("rp_IDJ", "rm_NameC", "rm_quantfund"))

MF_CT_nonalphafund(mfadd = NULL, mfrm = NULL)

MF_getAssetPort(fundID, rptDate, datasrc = "jy", TF)

MF_getTFSW(TF, fill = TRUE)
}
\arguments{
\item{fundtype}{\code{equity} means quity fund,\code{aggr_allo} means aggressive allocation fund.}

\item{type}{\code{rp_IDJ} means replace fundID's last four characters which is 'J.OF', \code{rm_NameC} means remove fundName last character which is C, \code{rm_quantfund} means remove quantative funds.}

\item{fill}{default value is \code{TRUE},if \code{TRUE} then fill quarter data with nearest half year report.}
}
\description{
\code{\bold{MF_getTF}} get mutual funds in RebDates.

\code{\bold{MF_TF_reshape}} reshape a \code{TF} object.

\code{\bold{MF_CT_nonalphafund}} export const non-alpha fundIDs.

MF_getAssetPort

\code{\bold{MF_getTFSW}} get TF with stockID and wgt
}
\examples{
RebDates <- getRebDates(as.Date('2010-01-01'),as.Date('2018-06-30'),'quarter')
TF <- MF_getTF(RebDates) #get all quity funds
TF <- MF_getTF(RebDates,fundtype='aggr_allo') #get all aggressive allocation funds
################~~MF_getTFSW demo~~###################################
TFSW <- MF_getTFSW(TF)
TFSW <- MF_getTFSW(TF,fill=FALSE)
}
