% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{queryAndClose.odbc}
\alias{queryAndClose.odbc}
\title{queryAndClose.odbc}
\usage{
queryAndClose.odbc(db, query, as.is = FALSE, ...)
}
\arguments{
\item{db}{a ODBC database object}

\item{query}{a character string,indicating the query to execute}
}
\value{
a dataframe
}
\description{
read data from a ODBC data source with a query
}
\examples{
queryAndClose.odbc(db.quant(),"select top 10 * from QT_DailyQuote")
}
\author{
Ruifei.Yin
}
