% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rdate2ts}
\alias{rdate2ts}
\title{rdate2ts}
\usage{
rdate2ts(rdate, GAP.R2TS = 25569)
}
\arguments{
\item{rdate}{a vector with \bold{Date} class}

\item{GAP.R2TS}{the constant of R to tinysoft date-transformating}
}
\value{
a vector containing the tinysoft date value transformed from the R date
}
\description{
transform the R date to tinysoft date value
}
\seealso{
\code{\link{tsdate2r}}
}
\author{
Ruifei.Yin
}
