% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{read.db.dbi}
\alias{read.db.dbi}
\title{read.db.dbi}
\usage{
read.db.dbi(db, query, ...)
}
\arguments{
\item{db}{a DBI data source object}

\item{query}{a character string,indicating the query to execute}
}
\value{
a dataframe
}
\description{
read data from a DBI data source with a query
}
\examples{
read.db.dbi(db.local(),"select * from QT_DailyQuote limit 10")
}
\author{
Ruifei.Yin
}

