% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{refinePar_default}
\alias{refinePar_default}
\title{refinePar_default}
\usage{
refinePar_default(type = c("none", "scale", "scale_sec", "reg", "reg_glm_sec",
  "reg_glm_cap"), sectorAttr = defaultSectorAttr("ind"), log = FALSE,
  regLists = list(fl_cap(log = TRUE, var = "float_cap", datasrc = "memory")))
}
\arguments{
\item{type}{Could be "none","scale","scale_sec","reg","reg_glm_sec","reg_glm_cap"}

\item{sectorAttr}{could be a sectorAttr list, or NULL(means no grouping by sector).}

\item{log}{TRUE or FALSE}

\item{regLists}{a \bold{factorLists}}
}
\description{
refinePar_default
}
\author{
Han.Qian
}
