% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{reg_mazi}
\alias{reg_mazi}
\title{regressing function for barra}
\usage{
reg_mazi(mtsfr_i, y, fnames, subset_reg = NULL)
}
\arguments{
\item{y}{string name for y.}

\item{fnames}{string vector for factor names.}

\item{subset_reg}{If not null, the column in mtsfr_i should include 'subset_reg_flag' to indicate the subset to regress.}
}
\value{
A list, including three values, frtn, res, est_rtn(estimated values).
}
\description{
the regression satisfies the specific one constriant optimization. mtsfr_i SHOULD WENT THROUGH  GF_CAP, GF_SECTOR before input.
}
