% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.deadline}
\alias{rptDate.deadline}
\title{rptDate.deadline
return the deadline of rptDate.}
\usage{
rptDate.deadline(rptDate)
}
\arguments{
\item{rptDate}{a vector of rptDate, with class Date.}
}
\value{
a vector of Date
}
\description{
rptDate.deadline
return the deadline of rptDate.
}
\examples{
rptDate.deadline(ymd(c(20100630,20141231,20130930)))
}
\author{
Ruifei.Yin
}
