% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.get}
\alias{rptDate.get}
\title{rptDate.yearrpt}
\usage{
rptDate.get(begT, endT, freq = c("q", "y", "h"), dir = 0L)
}
\arguments{
\item{freq}{"q","y","h"}

\item{dir}{1L,-1L,or 0L}
}
\value{
a vector of yearly rptDate.
}
\description{
rptDate.yearrpt
}
\examples{
rptDate.get(as.Date("2011-02-06"),as.Date("2011-10-23"),freq="h",dir=1)
rptDate.get(as.Date("2011-12-31"),as.Date("2013-06-30"),freq="q",dir=0)
}
\author{
Ruifei.Yin
}
