% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.is}
\alias{rptDate.is}
\title{rptDate.is}
\usage{
rptDate.is(rptDate, freq = c("q", "y", "h"))
}
\arguments{
\item{rptDate}{a vector of rptDate , with class Date.}
}
\value{
a logical vector with the same length of rptDate.
}
\description{
Is it a proper report date?
}
\examples{
rptDate.is(ymd(c(20100103,20141231,20130630)))
rptDate.is(ymd(c(20100630,20141231,20130930)))
rptDate.is(ymd(c(20100630,20141231,20130930)),"h")
}
\seealso{
\code{\link{check.rptDate}}
}
\author{
Ruifei.Yin
}
