% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.nearest}
\alias{rptDate.nearest}
\title{rptDate.yearrpt}
\usage{
rptDate.nearest(datelist, freq = c("q", "y", "h"), dir = -1L)
}
\arguments{
\item{datelist}{a vector of date , with class Date.}
}
\value{
a vector of yearly rptDate.
}
\description{
rptDate.yearrpt
}
\examples{
rptDate.nearest(as.Date(c("2015-01-11","2016-06-21","2016-12-31")),"q",1)
rptDate.nearest(as.Date(c("2015-01-11","2016-06-21","2016-12-31")),"y",-1)
}
\author{
Ruifei.Yin
}
