% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.offset}
\alias{rptDate.offset}
\title{rptDate.offset}
\usage{
rptDate.offset(rptDate, by, freq)
}
\arguments{
\item{rptDate}{a vector of rptDate , with class Date.}

\item{by}{a vector of integer}

\item{freq}{charactor string: "y" or "q".}
}
\value{
a vector or a dataframe according to the length of rptDate and by.
}
\description{
rptDate.offset
}
\examples{
# multi rptDate and 1 by:
rptDate.offset(as.Date(c("2016-03-31","2016-06-30","2016-09-30")),-1,"q")
rptDate.qoq(as.Date(c("2016-03-31","2016-06-30","2016-09-30")))
rptDate.offset(as.Date(c("2016-03-31","2016-06-30","2016-09-30")),-1,"y")
rptDate.yoy(as.Date(c("2016-03-31","2016-06-30","2016-09-30")))
# 1 rptDate and multi by:
rptDate.offset(as.Date("2016-03-31"),-2:0,"q")
# multi rptDate and multi by:
rptDate.offset(as.Date(c("2016-03-31","2016-06-30","2016-09-30")),-2:0,"q")
}
\author{
Ruifei.Yin
}
