% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.publ}
\alias{rptDate.publ}
\title{rptDate.publ}
\usage{
rptDate.publ(rptTS, rptDate, stockID, datasrc = defaultDataSRC())
}
\description{
rptDate.publ
}
\examples{
rptTS <- getrptTS(begT=as.Date("2016-02-06"),endT=as.Date("2017-8-23"),univ='EI000300')
re <- rptDate.publ(rptTS)
rptDate <- rptDate.get(as.Date("2011-02-06"),as.Date("2013-10-23"))
re <- rptDate.publ(rptDate=rptDate,stockID=c("EQ000001","EQ000002"))
}
\author{
Aming.Tao
}
