% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptDate.qoq}
\alias{rptDate.qoq}
\title{rptDate.qoq}
\usage{
rptDate.qoq(rptDate)
}
\arguments{
\item{rptDate}{a vector of rptDate , with class Date.}
}
\value{
a vector of qoq rptDate.
}
\description{
rptDate.qoq
}
\examples{
rptDate.qoq(lubridate::ymd(c(20100630,20141231,20130930)))
}
\author{
Ruifei.Yin
}
