% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptTS.getFin}
\alias{rptTS.getFin}
\alias{rptTS.getFin_ts}
\alias{rptTS.getFin_windR}
\title{rptTS.getFin}
\usage{
rptTS.getFin_windR(rptTS, field, ...)

rptTS.getFin_ts(rptTS, funchar, ...)
}
\arguments{
\item{rptTS}{a \bold{rptTS} object. a dataframe with cols:"rptDate","stockID"}

\item{field}{character strting or a vector of character string, giving the windfields. eg.  "OPEN,CLOSE,HIGH" or c("OPEN","CLOSE","HIGH")}

\item{...}{other arguments except \code{rptDate} in \code{w.wss}}

\item{funchar}{expression to get variables from tinysoft,a character string, usually copyed from tinysoft "stock-data-expert" and then replace the specified reportdate in the stock-data-expert expression by \code{'Rdate'}. e.g. convert \code{Last12MData(20091231,46002)} to \code{Last12MData(Rdate,46002)}.}

\item{...}{other arguments in \code{ts.wss}.}
}
\value{
a dataframe with the same length with rptTS, but added by some other financial indicator fields.
}
\description{
get financtial indicators via \bold{rptTS} through WindR, or TinySoft.
}
\examples{
TS <- Model.TS(modelPar.univ(indexID="ES09440000"))
TS <- getrptDate_newest(TS,mult="last")
rptTS <- unique(TS[, c("rptDate","stockID")])
# rptTS.getFin_windR
re <- rptTS.getFin_windR(rptTS,"np_belongto_parcomsh","rptType=1")
# rptTS.getFin_ts
re2 <- rptTS.getFin_ts(rptTS,'"np_belongto_parcomsh",report(46078,RDate)')
}
\author{
Ruifei.Yin
}
\seealso{
\code{\link[WindR]{w.wss}}

\code{\link{ts.wss}}
}

