% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{rptTS.getFinStat_ts}
\alias{rptTS.getFinStat_ts}
\title{rptTS.getFinStat_ts}
\usage{
rptTS.getFinStat_ts(rptTS, N, freq, funchar, varname, stat = c("mean", "sum",
  "slope", "slope/mean", "slope/growthmean", "sd", "mean/sd"), rm_N, ...)
}
\description{
rptTS.getFinStat_ts
}
\examples{
# rptTS.getFinStat_ts
FinStat <- rptTS.getFinStat_ts(rptTS,12,"q",'"np_belongto_parcomsh",report(46078,RDate)',stat="mean")
}
