% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{sector_NA_fill}
\alias{sector_NA_fill}
\title{deal with the NA value of sectorID}
\usage{
sector_NA_fill(sector, sectorAttr = defaultSectorAttr())
}
\arguments{
\item{TSS}{a dataframe with a "sector" colume}
}
\description{
replace the NA value of sectorID with an "OTHER" sector
}
