% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{sectorID2indexID}
\alias{sectorID2indexID}
\title{sectorID2indexID}
\usage{
sectorID2indexID(sectorID, std = 24)
}
\description{
sectorID2indexID
}
\examples{
sctID <- getSectorID(stockID = "EQ000001",drop=TRUE)
sectorID2indexID(sctID)
}
