% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{sectorNA_fill}
\alias{sectorNA_fill}
\title{deal with the NA value of sectorID}
\usage{
sectorNA_fill(TSS, sectorAttr = defaultSectorAttr())
}
\arguments{
\item{TSS}{a dataframe with a "sector" colume}
}
\description{
replace the NA value of sectorID with an "OTHER" sector
}

