% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{SecuMarket}
\alias{SecuMarket}
\title{SecuMarket}
\usage{
SecuMarket(stockID, datasrc = defaultDataSRC())
}
\arguments{
\item{stockID}{a vector}
}
\value{
a vector
}
\description{
SecuMarket
}
\seealso{
Other SecuMain functions: \code{\link{SecuCategory}},
  \code{\link{stockID2name}},
  \code{\link{stockID2stockID}},
  \code{\link{stockID2tradeCode}},
  \code{\link{stockName2ID}},
  \code{\link{tradeCode2stockID}}
}
