% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{setrefinePar}
\alias{setrefinePar}
\title{setrefinePar}
\usage{
setrefinePar(refinePar = refinePar_default(), outlier_method, outlier_par,
  outlier_sectorAttr, std_method, std_log, std_sectorAttr, std_regLists,
  std_regWgt, na_method, na_sectorAttr, all_sectorAttr)
}
\arguments{
\item{refinePar}{A refinePar list.}

\item{all_sectorAttr}{set all sectorAttr of std, outlier, na, in the same time}
}
\value{
A refinePar list.
}
\description{
setrefinePar
}
\author{
Han.Qian
}
