% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{stockID2indexID}
\alias{stockID2indexID}
\title{stockID2indexID}
\usage{
stockID2indexID(TS, stockID, withsector = FALSE)
}
\description{
stockID2indexID
}
\examples{
stockID2indexID(stockID = "EQ000001")
}
