% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{stockID2name}
\alias{stockID2name}
\title{stockID2name}
\usage{
stockID2name(stockID, IDsrc = "local", datasrc = defaultDataSRC())
}
\arguments{
\item{stockID}{a vector}
}
\value{
a vector
}
\description{
stockID2name
}
\examples{
stockID2name("EQ000527")
}
\seealso{
Other SecuMain functions: \code{\link{stockID2stockID}},
  \code{\link{stockID2tradeCode}},
  \code{\link{stockName2ID}},
  \code{\link{tradeCode2stockID}}
}

