% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{stockID2stockID}
\alias{stockID2stockID}
\title{stockID2stockID}
\usage{
stockID2stockID(stockID, from, to, datasrc = defaultDataSRC())
}
\arguments{
\item{stockID}{a vector of specific type}

\item{from}{a character string, giving the type of stockID convered from. Could be one of "local","ts","jy","wind".}

\item{to}{a character string, giving the type of stockID convered to. Could be one of "local","ts","jy","wind".}
}
\value{
a vector,return the stockID of specific datascr.
}
\description{
Covert the stockID to another type of stockID.
}
\examples{
stockID2stockID(c("EI000001","EI000300"),to="jy",from="local")
stockID2stockID(c("SH000001","SH000300"),to="jy",from="ts")
}
\author{
Ruifei.Yin
}
\seealso{
Other SecuMain functions: \code{\link{stockID2name}},
  \code{\link{stockID2tradeCode}},
  \code{\link{stockName2ID}},
  \code{\link{tradeCode2stockID}}
}

