% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{stockID2tradeCode}
\alias{stockID2tradeCode}
\title{stockID2tradeCode}
\usage{
stockID2tradeCode(stockID, IDsrc = "local", datasrc = defaultDataSRC())
}
\arguments{
\item{stockID}{a vector of stockID of specific datascr.}

\item{IDsrc}{a charactor string, could be one of "local","jy","ts","wind",etc.}
}
\value{
a vector of character,giving the stock trading code. eg. \code{c("600001","600002","000300")}
}
\description{
Covert the stockID to tradeCode.
}
\examples{
stockID2tradeCode(c("EI000001","EQ000030"),IDsrc="local")
stockID2tradeCode(c("SZ000001","SH000030"),IDsrc="ts")
}
\seealso{
Other SecuMain functions: \code{\link{SecuCategory}},
  \code{\link{SecuMarket}}, \code{\link{stockID2name}},
  \code{\link{stockID2stockID}},
  \code{\link{stockName2ID}},
  \code{\link{tradeCode2stockID}}
}
\author{
Ruifei.Yin
}
