% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{stockName2ID}
\alias{stockName2ID}
\title{stockName2ID}
\usage{
stockName2ID(name, datasrc = defaultDataSRC())
}
\arguments{
\item{name}{a characoter vector}
}
\value{
a dataframe with cols: 'SecuAbbr', 'ID'
}
\description{
stockName2ID
}
\examples{
stockName2ID("ST")
}
\seealso{
Other SecuMain functions: \code{\link{SecuCategory}},
  \code{\link{SecuMarket}}, \code{\link{stockID2name}},
  \code{\link{stockID2stockID}},
  \code{\link{stockID2tradeCode}},
  \code{\link{tradeCode2stockID}}
}

