% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{str_risk_fls_default}
\alias{str_risk_fls_default}
\title{default structure risk factor list}
\usage{
str_risk_fls_default()
}
\description{
This function is the default factor list setting which would be used for regression in sigma prediction.
}
