% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.count}
\alias{trday.count}
\title{trday.count}
\usage{
trday.count(begT = as.Date("1990-12-19"), endT = Sys.Date(),
  stockID = NULL)
}
\arguments{
\item{begT}{a Date object.The default is as.Date("1990-12-19")}

\item{endT}{a Date object.The default is Sys.Date()}

\item{stockID}{a character string or null. If null, the market trading days, other wise the trading days of specific stock.}
}
\value{
a integer
}
\description{
get the count of the trading days between \code{begT} and \code{endT}.
}
\examples{
trday.count(as.Date("2012-01-01"),as.Date("2012-09-30"))
trday.count(as.Date("2012-01-01"),as.Date("2012-09-30"),stockID="EQ000527")
}
\author{
Ruifei.Yin
}

