% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.get}
\alias{trday.get}
\title{trday.get}
\usage{
trday.get(begT = as.Date("1990-12-19"), endT = Sys.Date(), stockID = NULL,
  datasrc = "memory")
}
\arguments{
\item{begT}{a Date object.The default is as.Date("1990-12-19")}

\item{endT}{a Date object.The default is Sys.Date()}

\item{stockID}{a character string or null. If null, the market trading days, other wise the trading days of specific stock.}
}
\value{
a vector of class Date
}
\description{
get the trading date series from begT to endT
}
\examples{
re <- trday.get()
re <- trday.get(as.Date("2012-01-01"),as.Date("2012-09-30"))
re <- trday.get(as.Date("2012-07-01"),as.Date("2013-03-30"),stockID="EQ000527")
system.time(replicate(100,ii <- trday.get(datasrc="local"))) # 14.09
system.time(replicate(100,ii <- trday.get(datasrc="memory"))) # 0.39
system.time(replicate(100,ii <- trday.get(datasrc="local",stockID="EQ000527"))) # 16.38
system.time(replicate(100,ii <- trday.get(datasrc="memory",stockID="EQ000527"))) # 1.25
}
\author{
Ruifei.yin
}

