% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.IPO}
\alias{trday.IPO}
\title{trday.IPO}
\usage{
trday.IPO(stockID, datasrc = defaultDataSRC())
}
\arguments{
\item{stockID}{a vector}
}
\value{
a vector
}
\description{
trday.IPO
}
\seealso{
Other SecuMain functions: \code{\link{SecuCategory}},
  \code{\link{SecuMarket}}, \code{\link{stockID2name}},
  \code{\link{stockID2stockID}},
  \code{\link{stockID2tradeCode}},
  \code{\link{stockName2ID}},
  \code{\link{tradeCode2stockID}}
}
