% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.is}
\alias{trday.is}
\title{trday.is}
\usage{
trday.is(datelist, stockID = NULL, TS, drop)
}
\arguments{
\item{datelist}{a vector of class Date}

\item{stockID}{a character string or NULL If NULL, the market trading days, other wise the trading days of specific stock.}
}
\value{
a logical vecter with elements TRUE or FALSE depending on whether its argument is an tradingday
}
\description{
Identify if the date is trading day or not
}
\examples{
datelist <- seq(from=as.Date("2013-01-21"),to=as.Date("2013-01-30"),by="day")
trday.is(datelist)
trday.is(datelist,stockID="EQ000527")
trday.is(TS=TS)
}
\author{
Ruifei.Yin
}

