% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.last}
\alias{trday.last}
\title{trday.last}
\usage{
trday.last(datelist, stockID = NULL, TS, drop = FALSE)
}
\description{
get the last tradingday. If datelist is tradingday, return trday.nearby(by=-1), else, return trday.nearest(dir=-1)
}
\examples{
(datelist <- seq(from=as.Date("2013-01-21"),to=as.Date("2013-01-30"),by="day"))
trday.is(datelist)
# -- compare the following three output:
trday.last(datelist)
trday.nearest(datelist,dir=-1)
trday.nearby(datelist,by=-1)
# -- return TS
stockID=c("EQ000527","EQ000001")
trday.last(datelist = datelist,stockID=stockID)
}
