% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.nearby}
\alias{trday.nearby}
\title{trday.nearby}
\usage{
trday.nearby(datelist, by = 1, stockID = NULL)
}
\arguments{
\item{datelist}{a vector of class Date}

\item{by}{a integer giving the lagging days.If positive,get the earlyer tradingday,if negetive,get the later tradingday.}

\item{stockID}{a character string or null. If null, the market trading days, other wise the trading days of specific stock.}
}
\value{
a vector of trading days of class Date
}
\description{
get the nearby tradingday by shifting forward or backward.If the argument datelist is not a tradingday,the tradingday nearest by it will be firstly find by function \code{\link{trday.nearest}}.
}
\examples{
(datelist <- as.Date(c("2012-07-21","2012-07-22","2012-07-23","2013-01-30")))
trday.nearby(datelist,20) # the tradingday 20 days earlyer than datelist
trday.nearby(datelist,-20) # the tradingday 20 days later than datelist 
trday.nearby(datelist,-20,stockID="EQ000527") 
}
\author{
Ruifei.Yin
}

