% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.nearest}
\alias{trday.nearest}
\title{trday.nearest}
\usage{
trday.nearest(datelist, dir = -1L, stockID = NULL, TS, drop)
}
\arguments{
\item{datelist}{a vector of class Date}

\item{dir}{a integer. Indicating forward or backward to find, if \code{datelist} is not tradingday. -1 for forward, 1 for backward.}

\item{stockID}{a character string or null. If null, the market trading days, other wise the trading days of specific stock.}
}
\value{
a vector of class Date, the value of which is the nearest tradingday before/after the \code{datelist} if \code{datelist} is not a tradingday, otherwise,the \code{datelist} itself.
}
\description{
get the nearest tradingday.
}
\examples{
datelist <- as.Date(c("2012-07-21","2012-07-22","2012-07-23","2013-01-30"))
trday.nearest(datelist)
trday.nearest(datelist, dir = 1)
trday.nearest(datelist, dir = 1, stockID="EQ000527")
}
\author{
Ruifei.Yin
}

