% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{trday.offset}
\alias{trday.offset}
\title{trday.offset}
\usage{
trday.offset(datelist, by = months(1), stockID = NULL, dir = if
  (as.numeric(by) > 0) 1L else -1L, TS, drop)
}
\arguments{
\item{datelist}{a vector of class Date}

\item{by}{a period object. See detail in package \code{lubridate}.}

\item{stockID}{a character string or null. If null, the market trading days, other wise the trading days of specific stock.}
}
\value{
a vector of trading days of class Date
}
\description{
offset the datelist by months,quarters,ect. then get the nearest tradingday
}
\examples{
(datelist <- as.Date(c("2012-07-21","2012-07-22","2012-07-23","2013-01-30")))
trday.offset(datelist,months(1)) # the tradingdays 1 month after the datelist
trday.offset(datelist,months(-1)) # the tradingdays 1 month before the datelist
trday.offset(datelist,years(1))
trday.offset(datelist,months(-1),stockID="EQ000527") 
}
\author{
Ruifei.Yin
}
