% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{TS.getFin_ts}
\alias{TS.getFin_ts}
\title{TS.getFin_ts}
\usage{
TS.getFin_ts(TS, funchar, varname = funchar)
}
\arguments{
\item{TS}{a \bold{TS} object}

\item{funchar}{character string, giving a tinysoft expression to get the financtial indicators of the stock. The expression can be made simply by replaceing the specified reportdate in the stock-data-expert expression by \code{'Rdate'}. e.g. change \code{Last12MData(20091231,46002)} to \code{Last12MData(Rdate,46002)}.}

\item{varname}{character string, giving the name of the returned variables}
}
\value{
A \bold{TSF} object,a dataframe, containing at least cols:\bold{date}(with class of Date),\bold{stockID},\bold{varname},\bold{Rdate}
}
\description{
get financial factors throug a tinysoft expression, which could be simple or more complicated.
}
\details{
Note that the tinysoft function must contain ONLY two system parameters(no more!):pn_stock() and pn_date() ,which is supplyed from the two fields in the TS respectively.Also,the function must contain a expression of 'Rdate:=NewReportDateOfEndT2(sp_time())' to get the newest Rdate of sp_time.
}
\note{
Note that the tinysoft function must contain ONLY two system parameters(no more!):pn_stock() and pn_date() ,which is supplyed from the two fields in the TS respectively. Also,the function must contain a expression of 'Rdate:=NewReportDateOfEndT2(sp_time())' to get the newest Rdate of sp_time.
}
\examples{
TS <- getTS(getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17')),'EI000300')
TSF <- TS.getFin_ts(TS,"ReportOfAll(9900416,Rdate)")
TSF2 <- TS.getFin_ts(TS,"GrowthOfNReport(@LastQuarterData(DefaultRepID(),9900003,0),Rdate,3,1)")
TSF3 <- TS.getFin_ts(TS,"StockAveHsl2(20)+reportofall(9900003,Rdate)")
}
\author{
Ruifei.Yin
}

