% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{TS.getTech}
\alias{TS.getTech}
\title{TS.getTech}
\usage{
TS.getTech(TS, variables, exclude.TS = FALSE, tableName = "QT_DailyQuote",
  datasrc = defaultDataSRC())
}
\description{
TS.getTech
}

