% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{TS.sus_res}
\alias{TS.sus_res}
\title{TS.sus_res}
\usage{
TS.sus_res(TS, datasrc = defaultDataSRC())
}
\value{
a data.frame with cols: "sus" and "res"
}
\description{
get the suspension and resumption data
}
\author{
Ruifei.yin
}

