% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{TS_filter}
\alias{TS_filter}
\title{TS_filter}
\usage{
TS_filter(TS, begT, endT, sample_N)
}
\arguments{
\item{TS}{a TS, TSF, TSFR ...}

\item{sample_N}{integer. size of sample days}
}
\description{
sample or filter the TS by date.
}
\examples{
re <- TS_filter(ts,sample_N=8)
re <- TS_filter(ts,as.Date("2010-01-01"),as.Date("2011-01-01"))
}
