% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{TS_split}
\alias{TS_split}
\title{TS_split}
\usage{
TS_split(TS, by)
}
\arguments{
\item{by}{a vector, with the same length of TS}
}
\value{
a list is TS
}
\description{
split TS to lists, which are usually TSFs, TSFRs
}
\examples{
tsfr <- getSectorID(tsfr,sectorAttr = defaultSectorAttr("ind",336))
TSFRs1 <- TS_split(tsfr,by=tsfr$sector)
MC.chart.IC(tsfrs1)
TSFRs2 <- TS_split(tsfr,by=cut.Date2(tsfr$date,breaks = "2 year"))
MC.chart.IC(tsfrs2)
}
