% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{TSF_nextF}
\alias{TSF_nextF}
\title{TSF_nextF}
\usage{
TSF_nextF(TSF)
}
\value{
a \bold{TSFR}, in which the 'periodrtn' is the 'next_factorscore'.
}
\description{
\code{TSF_nextF} turn \code{TSF} into \code{TSFR} by lead date and factorscore one period.
}
\examples{
tsfr <- TSF_nextF(tsf)
# --- deal with the 'rptTSF'
rptTS <- getrptTS(as.Date("2011-12-31"),as.Date("2015-12-31"),freq = "q",univ = "EI000300")
rptTSF <- rptTS.getFin_windR(rptTS ,field = "np_belongto_parcomsh",varname = "factorscore")
TSF <- renameCol(rptTSF,"rptDate","date")
TSFR <- TSF_nextF(TSF)
# --- analysing
chart.IC(TSFR)
chart.Ngroup.overall(TSFR)
}
