% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{db.local}
\alias{db.local}
\alias{db.quant}
\alias{db.cs}
\alias{db.jy}
\alias{db.wind}
\title{Database connection}
\usage{
db.local(dbname = "QTlite", dbpath = "D:/sqlitedb")

db.quant()

db.cs()

db.jy()

db.wind()
}
\value{
a database connection.
}
\description{
connect database wind, quant, cs, jy, local, ...
}
