% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{defaultSectorAttr}
\alias{defaultSectorAttr}
\title{defaultSectorAttr}
\usage{
defaultSectorAttr()
}
\value{
a list, the value of the default sectorAttr
}
\description{
get the default sectorAttr. You can reset the default value by eg. \code{options(sectorAttr=list(std=3,level=2))}
}
\examples{
# -- get the default sectorAttr
defaultSectorAttr()
# -- reset 
options(sectorAttr=list(std=3,level=2))
# -- reget
defaultSectorAttr()
# -- combined sectorAttr
factorList1 <- buildFactorList(factorFun = "gf.mkt_cap", factorStd = "norm", factorNA = "na")
options(sectorAttr= list(std=list(factorList1,33),level=list(5,1)))
defaultSectorAttr()
}

