% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct01_frameBuildingFuncs.R
\name{factor_std}
\alias{factor_std}
\title{factor_std}
\usage{
factor_std(TSF, method = c("none", "scale", "robustScale", "reg"),
  log = FALSE, sectorAttr = defaultSectorAttr(), regLists = NULL)
}
\arguments{
\item{method}{Currently supporting four types : none, scale, robustScale, reg.}

\item{log}{Logical value. Default FALSE.}

\item{sectorAttr}{could be a sectorAttr list or NULL.}

\item{regLists}{factorLists. This argument will be put into use when method is reg. The factors in regLists will be stripped off from the factorscore.}
}
\description{
factor_std
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.NP_YOY(TS)
regLists <- buildFactorLists(buildFactorList(factorFun = 'gf.PE_ttm',factorDir = -1,factorRefine=refinePar_default("scale")))
TSF <- factor_std(TSF, method = "reg", log = FALSE, regLists = regLists)
}
\author{
Han.Qian
}
