% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk01_BARRA_Model.R
\name{get_frtn_res_series}
\alias{get_frtn_res_series}
\title{THE REGRESSION FUNCTION}
\usage{
get_frtn_res_series(ts, ts_union, barra_fls, barra_param)
}
\value{
A list, including three values, frtn_seri, res_seri, data_bank(mtsfr in fact, it would be passed into functions later to speed up the process.).
}
\description{
The subset_reg switch would not affect the number of observation in output, the res_seri would cover the whole ts.
}
