% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{getSectorID}
\alias{getSectorID}
\alias{gf_sector}
\alias{cast_sector}
\title{getSectorID}
\usage{
getSectorID(TS, stockID, endT = Sys.Date(),
  sectorAttr = defaultSectorAttr(), ret = c("ID", "name"), drop = FALSE,
  fillNA = FALSE, datasrc = defaultDataSRC())

gf_sector(TS, sectorAttr)

cast_sector(TSS)
}
\arguments{
\item{TS}{a \bold{TS} object}

\item{stockID}{a vector of stockID}

\item{endT}{a vector of Date}

\item{sectorAttr}{a list(See more in \code{\link{defaultSectorAttr}}) or NULL,or "existing".}

\item{ret}{a charactor string,could be one of "ID" or "name",indicating sectorID or sectorName returned.}

\item{drop}{a logical. Shoud the \code{TS} be exculded in the result?}

\item{TSS}{a dataframe, with cols: date,stockID,and some dummy variables of sectors.}
}
\value{
a data.frame,with the same cols of TS,added by "\code{sector}". Or a vector if \code{drop} is TRUE. You can get more sector infomation by \code{\link{CT_industryList}}
}
\description{
get the sectorID of the stocks on specific dates.
}
\note{
param TS and combination of stockID and endT should  at least have one and only have one. The combination of vector stockID and endT could be different length, which abide by the recycling rule.
}
\examples{
# - with TS
TS <- getTS(getRebDates(as.Date('2007-03-17'),as.Date('2012-05-20')),'EI000300')
getSectorID(TS)
# - one stock, multiple dates
getSectorID(stockID="EQ000001", endT=as.Date(c("2010-01-01","2012-01-01","2013-01-01")))
# - one date, multiple stocks
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"))
# - get 'ZHONGXIN' sector
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"), sectorAttr=list(3,1), ret="name")
getSectorID(stockID=c("EQ000001","EQ000002","EQ000004"), endT=as.Date("2010-01-01"), sectorAttr=list(3,2), ret="name")
# -- combined sectorAttr
test <- getSectorID(TS, sectorAttr= defaultSectorAttr("fct",fct_level = 5))
test <- getSectorID(TS, sectorAttr= defaultSectorAttr("ind_fct",ind_std = 33,ind_level = 1,fct_level = 5))
test <- getSectorID(TS, sectorAttr= defaultSectorAttr("ind_fct",ind_std = c(336,33),ind_level = 1,fct_level = 5))
test <- getSectorID(TS, sectorAttr= defaultSectorAttr("fct",fct_std = buildFactorLists_lcfs(c("F000001","F000006"),factorRefine = refinePar_default("none",NULL)),fct_level = 5))
factorList1 <- buildFactorList(factorFun = "gf_cap",factorRefine=refinePar_default("robust",NULL))
test2 <- getSectorID(TS, sectorAttr= list(std=list(factorList1,33),level=list(5,1)))
}
\author{
Ruifei.Yin
}
