% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{is_priceLimit}
\alias{is_priceLimit}
\title{is_priceLimit}
\usage{
is_priceLimit(TS, nearby = 0, lim = c(-10, 10), priceType = c("close",
  "open"), datelist, stockID, drop, datasrc = defaultDataSRC())
}
\arguments{
\item{nearby}{a integer vector. 0 means today, 1 means next tradingday. default is 0L. see detail in \code{\link{trday.nearby}}.}

\item{lim}{a vector of length 2.}

\item{priceType}{"close" or "open".}
}
\description{
if over-price-limit
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
re <- is_priceLimit(TS)
re1 <- is_priceLimit(TS,nearby=-1:1)
re2 <- is_priceLimit(TS,lim=c(-Inf,10)) #  limit-up
is_priceLimit(stockID = "EQ300576",datelist = as.Date("2016-12-22"),priceType = "open") #  open-price over limit
}
\author{
Ruifei.yin
}

