% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.add.LC_indexComponent}
\alias{lcdb.add.LC_indexComponent}
\title{add a index to local database from JY database}
\usage{
lcdb.add.LC_indexComponent(indexID)
}
\arguments{
\item{indexID}{is index code,such as EI000300}
}
\value{
nothing
}
\description{
add a index to local database from JY database
}
\examples{
lcdb.add.LC_indexComponent(indexID="EI801003")
lcdb.add.LC_indexComponent(indexID="EI000985")
}
\author{
Andrew Dow
}

