% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.init.indexquote000985}
\alias{lcdb.init.indexquote000985}
\title{lcdb.init.indexquote000985}
\usage{
lcdb.init.indexquote000985(windCode = "000985.CSI")
}
\description{
initiate the quote data of index 000985 through Wind API.
}
\examples{
lcdb.init.indexquote000985()
}

