% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.update.QT_FreeShares}
\alias{lcdb.update.QT_FreeShares}
\title{lcdb.build.QT_FreeShares}
\usage{
lcdb.update.QT_FreeShares(begT, endT, Freq = "week")
}
\description{
lcdb.build.QT_FreeShares
}
\examples{
lcdb.build.QT_FreeShares(filename="D:/sqlitedb/FREE_SHARES.csv")
lcdb.update.QT_FreeShares()
}
