% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{lcdb.update.QT_FreeShares}
\alias{lcdb.update.QT_FreeShares}
\title{lcdb.update.QT_FreeShares}
\usage{
lcdb.update.QT_FreeShares()
}
\description{
lcdb.update.QT_FreeShares
}
\examples{
library(WindR)
lcdb.update.QT_FreeShares()
}

