% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcdb.update}
\alias{lcdb.update}
\alias{lcdb.update.SecuMain}
\alias{lcdb.update.QT_TradingDay}
\alias{lcdb.update.CT_SystemConst}
\alias{lcdb.update.CT_IndustryList}
\alias{lcdb.update.LC_ExgIndustry}
\alias{lcdb.update.LC_IndexComponent}
\alias{lcdb.update.LC_IndexComponentsWeight}
\alias{indexID}
\alias{a}
\alias{vector}
\alias{of}
\alias{indexID.}
\alias{if}
\alias{missing,}
\alias{all}
\alias{the}
\alias{index}
\alias{in}
\alias{server}
\alias{database.}
\alias{lcdb.update.QT_IndexQuote}
\alias{lcdb.update.QT_DailyQuote}
\alias{lcdb.update.QT_DailyQuote2}
\alias{lcdb.update.LC_RptDate}
\alias{lcdb.update.LC_PerformanceGrowth}
\alias{lcdb.fix.swindustry}
\alias{lcdb.fix.ezindustry}
\title{update the local database}
\usage{
lcdb.update()

lcdb.update.SecuMain()

lcdb.update.QT_TradingDay()

lcdb.update.CT_SystemConst()

lcdb.update.CT_IndustryList()

lcdb.update.LC_ExgIndustry()

lcdb.update.LC_IndexComponent()

lcdb.update.LC_IndexComponentsWeight(begT, endT, IndexID)

lcdb.update.QT_IndexQuote(begT, endT, IndexID)

lcdb.update.QT_DailyQuote(begT, endT, stockID)

lcdb.update.QT_DailyQuote2(begT, endT, stockID)

lcdb.update.LC_RptDate(begT, endT)

lcdb.update.LC_PerformanceGrowth(begT, endT)

lcdb.fix.swindustry()

lcdb.fix.ezindustry()
}
\arguments{
\item{begT}{a numeric date. if missing, '\code{max(EndDate)}' in the lcdb.}

\item{endT}{a numeric date. if missing, 99990101.}
}
\value{
nothing.

nothing.
}
\description{
update the local database

Due to the SHENWAN inustry standard changing in 2014, fix local database's shenwan industry rule's bug and make the rule keep consistent. The new industry standard is 33.  Update 3 related local tables:CT_SystemConst,CT_IndustryList and LC_ExgIndusry.
}
\examples{
lcdb.update.LC_IndexComponentsWeight() # update all the indexs up to date
lcdb.update.LC_IndexComponentsWeight(20060101,20060330) # update all the indexs in given period
lcdb.update.LC_IndexComponentsWeight(19000101,99990101,"EI000905") # update all the data of given index
lcdb.update.LC_IndexComponentsWeight(20060101,20060330,"EI000905")
lcdb.fix.swindustry()
lcdb.fix.ezindustry()
}
\author{
Andrew Dow

Qian Han
}
