% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub03_DatabaseOperationFuncs.R
\name{lcfs.update}
\alias{lcfs.update}
\title{lcfs.update}
\usage{
lcfs.update(factorID, begT, endT, stockID, splitNbin = "month")
}
\arguments{
\item{factorID}{a single charactor of factorID}

\item{begT}{the begin date of the updating}

\item{endT}{the end date of the updating}

\item{stockID}{a vector of stockID}
}
\description{
update \bold{one} specific factorscore.
}
\examples{
# update a factorscore on all the time, of all the stocks
lcfs.update("F000008")
# update a factor on certain time
lcfs.update("F000008",20130322,20130330)
# update a factor of certain stocks
lcfs.update("F000008",20130322,20130330,c("EQ000001","EQ000002")) 
# update a factorscore on certin time, of certain stocks
lcfs.update("F000008",20130322,20130330,c("EQ000001","EQ000002"))
}
\seealso{
\code{\link{lcdb.update.QT_FactorScore}}, \code{\link{lcfs.add}}
}
