% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{MF_getQuote}
\alias{MF_getQuote}
\title{MF_getQuote}
\usage{
MF_getQuote(fundID, begT, endT, variables = "NAV_adj_return1",
  datasrc = c("wind", "jy", "ts"))
}
\description{
MF_getQuote
}
\examples{
fundID <- c("519983.OF","519987.OF")
begT <- as.Date("2016-01-05")
variables <- c("NAV_adj_return1")
MF_getQuote(fundID = fundID, begT = begT, variables = variables)
}
