% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{TS.getTech}
\alias{TS.getTech}
\title{TS.getTech}
\usage{
TS.getTech(TS, variables = select.list(CT_TechVars(datasrc = datasrc, secuCate
  = "EQ", tableName = tableName)[["varName"]], graphics = TRUE, multiple =
  TRUE), tableName = "QT_DailyQuote2", datasrc = defaultDataSRC())
}
\description{
TS.getTech
}

