% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub02_dataGettingFuncs.R
\name{tsInclude}
\alias{tsInclude}
\alias{tsLoad}
\alias{tsRequire}
\title{tsInclude}
\usage{
tsInclude(os = R.Version()$arch)

tsLoad(os = R.Version()$arch)

tsRequire()
}
\arguments{
\item{funchar}{a charactor string of tinysoft script}

\item{pars}{a list of funchar's parametres}

\item{syspars}{a list of tinysoft system parametres.(including:StockID CurrentDate Cycle bRate RateDay Precision)}
}
\value{
tsInclude: include and load the external dll;

tsRequre: test if tinysoft connected, if not, connect it.
}
\description{
Series of functions of connect/login Tinysoft server, and remote call funcions of Tinysoft.
}
\examples{
tsInclude()
tsConnect()
tsLogined()
tsRemoteExecute("return close();",list(StockID="SZ000002",CurrentDate=rdate2ts(as.Date("2014-11-11"))))
# use ldply/laply/llply to transform the result from list to comfortable form
ll <- tsRemoteExecute("return Array(('a':1,'b':4),('a':2,'b':5),('a':3,'b':6));")
plyr::ldply(ll,as.data.frame)
plyr::laply(ll,as.array)
plyr::llply(ll,unlist)
tsRemoteCallFunc("close",,list(StockID="SZ000002"))
tsRemoteCallFunc("rand",list(2,3))
tsDisconnect()
}
\author{
ruifei.yin
}
